requirements.txt
python-telegram-bot
asyncio==3.4.3
SQLAlchemy==1.4.49
Pillow==9.5.0
psutil==5.9.8
pretty_errors==1.2.25
pyjokes
psycopg2-binary==2.9.9
pynewtonmath==1.0.1
pyrate-limiter==3.6.0
pytz==2023.3.post1
python-arq==6.0.7
pyrogram==2.0.106